requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
tweepy==3.10.0
numpy==1.22.3
scikit-learn==1.0.2
pandas>=0.19
mecab-python3 == 1.0.4
unidic-lite==1.0.8
nltk==3.7
scipy==1.8.0